+++ .gitignore
@@ -8,3 +8,8 @@ config.json
 /dist
 /get_srs_data
 /src/backup
+src/Field.class.js
+src/Field.js
+src/Mino.class.js
+src/Mino.js
+src/Wetris.class.js
